Renaming the project to crmpl
Papr is taking a different direction to focus more into readability giving up on a data format with small foot print. This format still has its usefulness as it attempts to solve a different problem. So I'm not deleting this. And since, this format is like crushing a paper into a ball, I'm going to call this one crmpl (like crumpling a paper).